+++ src/main/kotlin/lotto/utils/InputView.kt
@@ -38,10 +38,10 @@ object InputView {
             throw IllegalArgumentException("[ERROR] The purchase amount must be a multiple of 1000.")
     }
 
-    fun toWinningNumbers(str: String): List<UByte> {
+    fun toWinningNumbers(str: String): List<Int> {
         val winningNumbers: List<String> = str.split(",")
         return try {
-            winningNumbers.map { it.toUByte() }
+            winningNumbers.map { it.toInt() }
         } catch (e: NumberFormatException) {
             throw IllegalArgumentException("[ERROR] Invalid input for winning numbers.")
         }
